requirements.txt
click==8.0.1
azure-mgmt-resource==20.0.0
azure-identity==1.6.1
pytest==6.2.5
pytest-click==1.0.2
ruamel.yaml==0.17.16